requirements.txt
flask
python-dotenv
google-search-results
scikit-learn
requests
